Refactor manual attachment registries
This commit refactors the attachment service registries:

- renaming the `ManualDocumentAttachmentServiceRegistry` to
  `OrganisationalBlahBlah...` to reflect that it is scoped to an
  Organisation

- introduced a service registry that isn't scoped to an organisation:
  `ManualDocumentAttachmentServiceRegistry`

- moved shared code into an Abstract parent class for both the
  registries so that they only need to define which repositories they
  use
